feat (Phone/au):  Add in country:au for phone number generator.  Split phone tests out to separate file.
